refactor(wiki): extract handleWikiArticle to keep wiki_worker.go under the LOC budget
The attached-artifacts change pushed wiki_worker.go to 1511 LOC, over the
repo's 1500 hard limit (scripts/check-file-size.sh). Move the
GET /wiki/article handler verbatim into wiki_worker_article_handler.go
(same package, identical behavior). wiki_worker.go is now 1444 LOC.
